+++ packages/myst-cli/src/session/session.ts
@@ -25,12 +25,12 @@ import { rootReducer } from '../store/reducers.js';
 import version from '../version.js';
 import type { ISession } from './types.js';
 import { isWhiteLabelled } from '../utils/whiteLabelling.js';
-import { KernelManager, ServerConnection, SessionManager } from '@jupyterlab/services';
-import type { JupyterServerSettings } from 'myst-execute';
-import { launchJupyterServer } from 'myst-execute';
+import { SessionManager } from '@jupyterlab/services';
+import { createJupyterSessionManager } from 'myst-execute';
 import type { RequestInfo, RequestInit } from 'node-fetch';
 import { default as nodeFetch, Headers, Request, Response } from 'node-fetch';
 import type { PluginInfo } from 'myst-config';
+import { PluginRegistry } from '@lumino/coreutils';
 
 // fetch polyfill for node<18
 if (!globalThis.fetch) {
@@ -100,6 +100,8 @@ export class Session implements ISession {
     return this.$logger;
   }
 
+  private readonly registry: PluginRegistry<ISession>;
+
   constructor(
     opts: {
       logger?: Logger;
@@ -127,6 +129,8 @@ export class Session implements ISession {
         this._latestVersion = latest;
       })
       .catch(() => null);
+    this.registry = new PluginRegistry();
+    this.registry.application = this;
   }
 
   showUpgradeNotice() {
@@ -239,41 +243,9 @@ export class Session implements ISession {
 
   jupyterSessionManager(): Promise<SessionManager | undefined> {
     if (this._jupyterSessionManagerPromise === undefined) {
-      this._jupyterSessionManagerPromise = this.createJupyterSessionManager();
+      this._jupyterSessionManagerPromise = createJupyterSessionManager(this);
     }
     return this._jupyterSessionManagerPromise;
-  }
-
-  private async createJupyterSessionManager(): Promise<SessionManager | undefined> {
-    try {
-      let partialServerSettings: JupyterServerSettings | undefined;
-      // Load from environment
-      if (process.env.JUPYTER_BASE_URL !== undefined) {
-        partialServerSettings = {
-          baseUrl: process.env.JUPYTER_BASE_URL,
-          token: process.env.JUPYTER_TOKEN,
-        };
-      } else {
-        // Note: To use an existing Jupyter server use `findExistingJupyterServer`, see #1716
-        this.log.debug(`Launching jupyter server on ${this.sourcePath()}`);
-        // Create and load new server
-        partialServerSettings = await launchJupyterServer(this.sourcePath(), this.log);
-      }
-
-      const serverSettings = ServerConnection.makeSettings(partialServerSettings);
-      const kernelManager = new KernelManager({ serverSettings });
-      const manager = new SessionManager({ kernelManager, serverSettings });
-
-      // Tie the lifetime of the kernelManager and (potential) spawned server to the manager
-      manager.disposed.connect(() => {
-        kernelManager.dispose();
-        partialServerSettings?.dispose?.();
-      });
-      return manager;
-    } catch (err) {
-      this.log.error('Unable to instantiate connection to Jupyter Server', err);
-      return undefined;
-    }
   }
 
   dispose() {
